Delay get_data(scope) calls until there has been a trigger event (#228)
Replaced `stop(scope)` with `write(scope, 'digitize')`; a blocking
command that fills the internal buffer of the scope, therefore greatly
reducing the possibility of a timeout error/unsynced buffer. After the
buffer is filled, the scope automatically stops and allows for any other
command to be called, ex: `waveform:data?`.